修改: 1.依赖tendermint版本更新为0.27.0-dev1       2.不能为负数的字段均使用uint64类型

+++ client/qcp/cmd.go
@@ -102,7 +102,7 @@ $ basecli qcp tx [chainID] --seq [Seq]
 			seq := viper.GetInt64(flagOutSeq)
 
 			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			result, err := GetGetOutChainTx(cliCtx, outChainID, seq)
+			result, err := GetGetOutChainTx(cliCtx, outChainID, uint64(seq))
 
 			if err != nil {
 				return err
